services/user-service/requirements.txt
Flask>=2.0
PyMySQL
SQLAlchemy
alembic
passlib[bcrypt]
PyJWT
marshmallow
bcrypt